requirements.txt
codeclimate-test-reporter==0.1.1
coverage==4.1
coveralls==1.1
dj-database-url==0.4.1
Django==1.9.7
django-dotenv==1.4.1
django-getenv==1.3.1
docopt==0.6.2
factory-boy==2.7.0
fake-factory==0.5.7
gunicorn==19.6.0
ipaddress==1.0.16
psycopg2==2.6.1
python-dateutil==2.5.3
requests==2.10.0
six==1.10.0
whitenoise==3.2
